==> notebook/1.2_eda.ipynb <==
print("Hello, world")

!pip install numpy
!pip install --use-pep517 faiss-cpu
!pip install --use-pep517 faiss-gpu
!pip install fastparquet
!pip install --user matplotlib
!pip install phik

import pandas as pd
import numpy as np
import faiss
from tqdm.notebook import tqdm
from sklearn.preprocessing import RobustScaler

import phik
from phik.report import plot_correlation_matrix
from phik import report

import matplotlib.pyplot as plt

#from joblib.parallel import parallel_config
from joblib import parallel

import os
os.environ['JOBLIB_TEMP_FOLDER'] = 'F:/tmp/'


## Load data

### Base data

dict_base = {}
for i in range(72):
    dict_base[str(i)] = 'float32'
dict_base
dict_train = dict_base.copy()
dict_train['Target'] = 'str'

df_base = pd.read_csv("../data/base.csv", index_col=0, dtype=dict_base)
df_base.head()

### Train data

df_train = pd.read_csv("../data/train.csv", index_col=0, dtype=dict_train)
df_train.shape

df_train.head()

df_validation = pd.read_csv("../data/validation.csv", index_col=0)
df_validation.shape

df_validation.head()

df_validation_answer = pd.read_csv("../data/validation_answer.csv", index_col=0)
df_validation_answer.shape

df_validation_answer.head()

## EDA

df_base.hist(figsize=[20, 20], bins=50);

strange_features = ["6", "21", "25", "33", "44", "59", "65", "70"]

df_base[strange_features].hist(figsize=[15, 15], bins=100)


Фичи 6, 59,65 выглядят интересно в плане распределения. Возможно, данные фичи связаны с какими-то стандартными величинами, типа размера упаковки или чего-то подобного.
Несмотря на то, что у 6 фичи есть явный выброс, ее можно оставить и создать новую категориальную фичу.
59 и 65 - так же выглядят, как кандидаты для категоризации.

Фичи 33 и 44 - распеределение не особо радует, поэтому категоризурем данные фичи.

Фичи 21, 25, 70 - по идее, тоже можно было бы категоризовать, но у них есть сильный выбросы, что скорее всего помешает обучению.
Данные фичи нужно исключить из выборки.

Итого. 21, 25, 70 - удаляем из выборке.
6, 59, 65,33 и 44 - сотавляем для дальнейшего анализа.

df_base = df_base.drop(columns=["21", "25", "70"])
df_train = df_train.drop(columns=["21", "25", "70"])

### Phik Correlations

phik_corrs = df_base.phik_matrix()

%%time
plot_correlation_matrix(phik_corrs.values, 
                        x_labels=phik_corrs.columns, 
                        y_labels=phik_corrs.index, 
                        vmin=0, vmax=1, color_map="Blues", 
                        title=r"correlation $\phi_K$", 
                        fontsize_factor=1.5, 
                        figsize=(60, 60))
plt.show()

Отрицательный результат, тоже результат.
По большоей части, ни каких корреляция между фичами нет, а значит и сильной мультиколлинеарности тоже не будет.
Оставяем данные, как есть.

## Save data

df_base.to_parquet("../tmp/cleared_df_base.parquet", compression="gzip")

df_train.to_parquet("../tmp/cleared_df_train.parquet", compression="gzip")
