==> notebook/3_model_usearch.ipynb <==
print("Hello, world")

!pip install numpy usearch

import sys  
sys.path.insert(1, '../src')
import recsys_support.support_func as rs

import numpy as np
import pandas as pd
from usearch.index import Index
from sklearn.preprocessing import RobustScaler

index = Index(ndim=3)

vector = np.array([0.2, 0.6, 0.4])
index.add(42, vector)

matches = index.search(vector, 10)
assert matches[0].key == 42
assert matches[0].distance <= 0.001
#assert np.allclose(index[42], vector)

### Load Data

df_base = pd.read_parquet("../tmp/cleared_df_base.parquet")

df_train = pd.read_parquet("../tmp/cleared_df_train.parquet")

def checkRobustScaler(base, train):
    scaler = RobustScaler()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])

    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    return tbase, ttrain, scaler

df_base, df_train, scaler = checkRobustScaler(df_base, df_train)

df_base = df_base.reset_index()
df_train = df_train.reset_index()

df_base_ids = df_base.pop("Id")
df_train_ids = df_train.pop("Id")
df_train_target = df_train.pop("Target")

df_base.head(5)

df_train.head(5)

### Test Index

dims = df_base.shape[1]

index = Index(
    ndim=dims, # Define the number of dimensions in input vectors
    metric='inner_product', # Choose 'l2sq', 'haversine' or other metric, default = 'ip'
    dtype='f32', # Quantize to 'f16' or 'i8' if needed, default = 'f32'
    #connectivity=16, # Optional: Limit number of neighbors per graph node
    #expansion_add=128, # Optional: Control the recall of indexing
    #expansion_search=32, # Optional: Control the quality of the search
    multi=False, # Optional: Allow multiple vectors per key, default = False
)
#"cos": MetricKind.Cos,
#"cosine": MetricKind.Cos,
#ip": MetricKind.IP,
#"dot": MetricKind.IP,
#inner_product": MetricKind.IP,
#"l2sq": MetricKind.L2sq,
#"l2_sq": MetricKind.L2sq,
#"haversine": MetricKind.Haversine,
#"divergence": MetricKind.Divergence,
#"pearson": MetricKind.Pearson,
#"hamming": MetricKind.Hamming,
#"tanimoto": MetricKind.Tanimoto,
#"sorensen": MetricKind.Sorensen,


%%time
index.add(df_base.index, df_base.values)

%%time
matches = index.search(df_train.values, 10000)

%%time
predict_ids = []
for i in matches:
    predict_ids.append([z.key for z in i])

# функция вычисления accuracy@n
def accuracy(targets, predict_ids, base_index):
    acc = 0
    for target, candidates in zip(targets.tolist(), predict_ids):
        acc += int(target in [base_index[number] for number in candidates])
    print(acc)
    return (acc / len(predict_ids) * 100)

%%time
accuracy(df_train_target, predict_ids, {k: df_base_ids.iloc[v] for k, v in enumerate(df_base_ids.index.to_list())})
