==> notebook/6_faiss_with_important_features-and-normalization.ipynb <==
print("Hello, world")

import pandas as pd
import numpy as np
import faiss

from tqdm.notebook import tqdm
from sklearn.preprocessing import RobustScaler
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import Normalizer
from sklearn.preprocessing import MaxAbsScaler

important_feautres = ['20', 'norm', 'log_20', '63_minus_10', '10_minus_63', '63', '10',
       '63_minus_35', 'log_33', '35_plus_63', '46', '64', '61', '49',
       '55', '35', '12', '8', '36', '39', '24', '26', '29', '60', '32',
       '28', '27', '15', '52', '19', '37', '68', '22', '67', '3', '66',
       '47', '30', '18', '56', '4', '50', '48', '11', '45', '58', '62',
       '9', '34', '0', '41', '14', '17', '31', '71', '69', '43', '57',
       '40', '53', '51', '16', '38']

## Head-on Solution

df_base = pd.read_parquet("../tmp/cleared_df_base_new_features.parquet")

df_base.head()

df_train = pd.read_parquet("../tmp/cleared_df_train_new_features.parquet")
df_train.head()

dims = df_base.shape[1]
n_cells = 50
# how many cells
quantizer = faiss.IndexFlatL2(dims)
idx_l2 = faiss.IndexIVFFlat(quantizer, dims, n_cells)

idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))

base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
#del df_base

headon_solution_df_train = df_train.copy()
targets = headon_solution_df_train["Target"]
headon_solution_df_train.drop("Target", axis=1, inplace=True)

%%time
k=5
vecs, idx = idx_l2.search(np.ascontiguousarray(headon_solution_df_train.values).astype('float32'), k)

acc = 0
for target, el in zip(targets.values.tolist(), idx.tolist()):
    acc += int(target in [base_index[r] for r in el])

print(100 * acc / len(idx))

### Вывод
Решение в лоб - 5.575 попугаев.
Очень низкая оценка.

## Scalers

### StandardScaler

def checkStandardScaler(base, train):
    scaler = StandardScaler()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])

    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    return tbase, ttrain, scaler

sscaler_base, sscaler_train, _ = checkStandardScaler(df_base, df_train)

def buildA5Score(df_base, df_train, short=True, k=10):
    dims = df_base.shape[1]
    n_cells = 50
    # how many cells
    quantizer = faiss.IndexFlatL2(dims)
    idx_l2 = faiss.IndexIVFFlat(quantizer, dims, n_cells)
    idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
    idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))
    
    base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
    #del df_base
    
    headon_solution_df_train = df_train.copy()
    targets = headon_solution_df_train["Target"]
    headon_solution_df_train.drop("Target", axis=1, inplace=True)

    vecs, idx = idx_l2.search(np.ascontiguousarray(headon_solution_df_train.values).astype('float32'), k)
    
    acc = 0
    for target, el in zip(targets.values.tolist(), idx.tolist()):
        acc += int(target in [base_index[r] for r in el])
    
    return 100 * acc / len(idx)

%%time
print("A@5 Score on Standard Scaler :", buildA5Score(sscaler_base, sscaler_train))

### RobustScaler

def checkRobustScaler(base, train):
    scaler = RobustScaler()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])

    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    return tbase, ttrain, scaler

important_feautres_for_train = important_feautres.copy()
important_feautres_for_train.append("Target")
rscaler_base, rscaler_train, _ = checkRobustScaler(df_base[important_feautres], df_train[important_feautres_for_train])

%%time
print("A@5 Score on Robust Scaler :", buildA5Score(rscaler_base, rscaler_train, k=5))

### MinMaxScaler

def checkMinMaxScaler(base, train):
    scaler = MinMaxScaler()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])

    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    return tbase, ttrain, scaler

mmscaler_base, mmscaler_train, _ = checkMinMaxScaler(df_base, df_train)

%%time
print("A@5 Score on Min Max Scaler :", buildA5Score(mmscaler_base, mmscaler_train))

### Normalizer

def checkNormalizer(base, train):
    scaler = Normalizer()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])

    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    return tbase, ttrain, scaler

normscaler_base, normscaler_train, _ = checkNormalizer(df_base, df_train)

%%time
print("A@5 Score on Normalizer :", buildA5Score(normscaler_base, normscaler_train))

### MaxAbsScaler

def checkMaxAbsScaler(base, train):
    scaler = MaxAbsScaler()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])

    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    return tbase, ttrain, scaler

maxAbsscaler_base, maxAbsscaler_train, _ = checkMaxAbsScaler(df_base, df_train)

%%time
print("A@5 Score on Max Abs Scaler :", buildA5Score(maxAbsscaler_base, maxAbsscaler_train))

## FIASS Tuning

### IndexLSH

lsh_failed_targets = []

def buildA5ScoreV2(df_base, df_train, short=True, k_value=50):
    dims = df_base.shape[1]
    n_cells = 100
    # how many cells
    #quantizer = faiss.IndexHNSWFlat(dims)
    #idx_l2 = faiss.IndexIVFFlat(quantizer, dims, n_cells)
    idx_l2 = faiss.IndexLSH(dims, 2048)
    if short:
        idx_l2.train(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
    else:
        idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))
    
    base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
    #del df_base
    
    headon_solution_df_train = df_train.copy()
    targets = headon_solution_df_train["Target"]
    headon_solution_df_train.drop("Target", axis=1, inplace=True)

    k=k_value
    vecs, idx = idx_l2.search(np.ascontiguousarray(headon_solution_df_train.values).astype('float32'), k)
    
    acc = 0
    for target, el in zip(targets.values.tolist(), idx.tolist()):
        plus_one = int(target in [base_index[r] for r in el])
        acc += plus_one
        if plus_one == 0:
            lsh_failed_targets.append(target)
    
    return 100 * acc / len(idx), idx_l2

important_feautres = ['20', 'norm', 'log_20', '63_minus_10', '10_minus_63', '63', '10',
       '63_minus_35', 'log_33', '35_plus_63', '46', '64', '61', '49',
       '55', '35', '12', '8', '36', '39', '24', '26', '29', '60', '32',
       '28', '27', '15', '52', '19', '37', '68', '22', '67', '3', '66',
       '47', '30', '18', '56', '4', '50', '48', '11', '45', '58', '62',
       '9', '34', '0', '41', '14', '17', '31', '71', '69', '43', '57',
       '40', '53', '51', '16', '38']

important_feautres_for_train = important_feautres.copy()
important_feautres_for_train.append("Target")

#numer_of_imprtant_features = len(important_feautres)
#mat = faiss.PCAMatrix (numer_of_imprtant_features, numer_of_imprtant_features-10)
#mat.train(df_base[important_feautres].values)
#assert mat.is_trained
#base_array = mat.apply(df_base[important_feautres].values)
#train_array = mat.apply(df_train[important_feautres].values)
# print this to show that the magnitude of tr's columns is decreasin
#df_pca_base = pd.DataFrame(base_array)
#df_pca_train = pd.DataFrame(train_array)
#df_pca_train["Target"] = pd.Series(df_train["Target"].values)


rscaler_base, rscaler_train, _ = checkRobustScaler(df_base[important_feautres], df_train[important_feautres_for_train])

%%time
lsh_score, lsh_index = buildA5ScoreV2(rscaler_base, rscaler_train, short=False, k_value=5)
print("A@5 Score on Robust Scaler V2:", lsh_score)

lsh_score

lsh_failed_targets[0]

train1 = rscaler_train[rscaler_train['Target'] == '675816-base']

base1 = rscaler_base[rscaler_base.index == '675816-base']
base1

base2 = rscaler_base[rscaler_base.index == '755584-base']
base2

test1 = train1.copy()
test1.pop("Target")
np.linalg.norm(test1.values)

test2 = base1.copy()
np.linalg.norm(test2.values)

test3 = base2.copy()
np.linalg.norm(test3.values)

#search_data = rscaler_train[rscaler_train['Target'] == '675816-base'];
vecs, idx = lsh_index.search(np.ascontiguousarray(test1.values).astype('float32'), 100)
idx

### IndexHNSWFlat

def buildA5ScoreV3(df_base, df_train, short=True, k_value=5):
    dims = df_base.shape[1]
    n_cells = 100
    # how many cells
    #quantizer = faiss.IndexFlatL2(dims)
    #idx_l2 = faiss.IndexIVFFlat(quantizer, dims, n_cells)
    #idx_l2 = faiss.IndexLSH(dims, 1024)
    #M = 2
    #d2 = int((dims + M - 1) / M) * M
    #q = faiss.IndexPQ (d2, M, 4)
    #index_pq = faiss.IndexRefineFlat(q)
    #remapper = faiss.RemapDimensionsTransform (dims, d2, True)
    #idx_l2 = faiss.IndexPreTransform (remapper, index_pq)
    #idx_l2 = faiss.index_factory(dims, "OPQ16_64,IMI2x8,PQ8+16")

    idx_l2 = index = faiss.IndexHNSWFlat(dims, 512, faiss.METRIC_JensenShannon)
    
    if short:
        idx_l2.train(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
    else :
        idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))
    
    base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
    #del df_base
    
    headon_solution_df_train = df_train.copy()
    targets = headon_solution_df_train["Target"]
    headon_solution_df_train.drop("Target", axis=1, inplace=True)

    k=k_value
    vecs, idx = idx_l2.search(np.ascontiguousarray(headon_solution_df_train.values).astype('float32'), k)
    acc = 0
    for target, el in zip(targets.values.tolist(), idx.tolist()):
        acc += int(target in [base_index[r] if r in base_index else 0 for r in el])
    
    return 100 * acc / len(idx)



%%time
print("A@5 Score on Robust Scaler V3:", buildA5ScoreV3(rscaler_base, rscaler_train, short=False, k_value=5))

def buildA5ScoreV4(df_base, df_train, short=True, k_value=5):
    dims = df_base.shape[1]
    n_cells = 100
    # how many cells
    quantizer = faiss.IndexFlatL2(dims)
    idx_l2 = faiss.IndexIVFFlat(quantizer, dims, n_cells)

    if short:
        idx_l2.train(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
    else :
        idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))
    
    base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
    #del df_base
    
    headon_solution_df_train = df_train.copy()
    targets = headon_solution_df_train["Target"]
    headon_solution_df_train.drop("Target", axis=1, inplace=True)

    k=5
    vecs, idx = idx_l2.search(np.ascontiguousarray(headon_solution_df_train.values).astype('float32'), k)
    acc = 0
    for target, el in zip(targets.values.tolist(), idx.tolist()):
        acc += int(target in [base_index[r] if r in base_index else 0 for r in el])
    
    return 100 * acc / len(idx)
    

%%time
print("A@5 Score on Robust Scaler V3:", buildA5ScoreV4(rscaler_base, rscaler_train, short=False, k_value=5))

def buildA5ScoreV5(df_base, df_train, short=True, k_value=5):
    dims = df_base.shape[1]
    n_cells = 100
    # how many cells
    idx_l2 = faiss.index_factory(dims, "IVF69,SQ8")


    if short:
        idx_l2.train(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base[:50000].values).astype('float32'))
    else :
        idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))

    base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
    #del df_base
    
    headon_solution_df_train = df_train.copy()
    targets = headon_solution_df_train["Target"]
    headon_solution_df_train.drop("Target", axis=1, inplace=True)

    k=k_value
    vecs, idx = idx_l2.search(np.ascontiguousarray(headon_solution_df_train.values).astype('float32'), k)
    acc = 0
    for target, el in zip(targets.values.tolist(), idx.tolist()):
        acc += int(target in [base_index[r] if r in base_index else 0 for r in el])
    
    return 100 * acc / len(idx)
    

%%time
print("A@5 Score on Robust Scaler V3:", buildA5ScoreV5(rscaler_base, rscaler_train, short=False))
