==> notebook/7_faiss_with_important_features-and-new-features.ipynb <==
print("Hello, world")

!pip install tqdm-multiprocess

import pandas as pd
import numpy as np
import faiss
import time

from tqdm.notebook import tqdm
from tqdm import trange
from sklearn.preprocessing import RobustScaler
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import Normalizer
from sklearn.preprocessing import MaxAbsScaler

from sklearn.metrics.pairwise import additive_chi2_kernel
from sklearn.metrics.pairwise import chi2_kernel
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.metrics.pairwise import cosine_distances
from sklearn.metrics.pairwise import euclidean_distances
from sklearn.metrics.pairwise import haversine_distances
from sklearn.metrics.pairwise import laplacian_kernel
from sklearn.metrics.pairwise import linear_kernel
from sklearn.metrics.pairwise import manhattan_distances
from sklearn.metrics.pairwise import nan_euclidean_distances
from sklearn.metrics.pairwise import pairwise_kernels
from sklearn.metrics.pairwise import polynomial_kernel
from sklearn.metrics.pairwise import rbf_kernel
from sklearn.metrics.pairwise import sigmoid_kernel
from sklearn.metrics.pairwise import paired_euclidean_distances
from sklearn.metrics.pairwise import paired_manhattan_distances
from sklearn.metrics.pairwise import paired_cosine_distances
from sklearn.metrics.pairwise import paired_distances

import multiprocessing
from multiprocessing import Pool
from tqdm.contrib.concurrent import process_map

## Head-on Solution

df_base = pd.read_parquet("../tmp/cleared_df_base_new_features.parquet")
df_train = pd.read_parquet("../tmp/cleared_df_train_new_features.parquet")
df_validation = pd.read_parquet("../tmp/cleared_df_validation_new_features.parquet")

df_base.head()

df_train["Target"].head()

## A-n Score function

def getAScore(targets, seached_idxs, base_index):
    defineActualBaseTargets = lambda x: base_index[x]
    
    start_score = time.time()
    acc = 0
    for target, el in zip(targets, seached_idxs):
        acc += int(np.isin(target, np.array([base_index[r] for r in el])))
    print("Score time:", time.time() - start_score)
    return 100 * acc / len(seached_idxs)

## Scalers

def buildA5Score(df_base, df_train, short=True, k=10):
    dims = df_base.shape[1]
    n_cells = 50
    # how many cells
    quantizer = faiss.IndexFlatL2(dims)
    idx_l2 = faiss.IndexIVFFlat(quantizer, dims, n_cells)
    idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
    idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))
    
    base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
    #del df_base
    
    headon_solution_df_train = df_train.copy()
    targets = headon_solution_df_train["Target"]
    headon_solution_df_train.drop("Target", axis=1, inplace=True)

    vecs, idx = idx_l2.search(np.ascontiguousarray(headon_solution_df_train.values).astype('float32'), k)
    
    score = getAScore(targes.values.tolist(), idx.to_list(), base_index)
    return score

### RobustScaler

def checkRobustScaler(base, train, validation):
    scaler = RobustScaler()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])
    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    tvalidation = scaler.transform(validation[base.columns])
    tvalidation = pd.DataFrame(tvalidation, columns=base.columns)
    tvalidation["Id"] = tvalidation.index
    tvalidation = tvalidation.set_index(["Id"])

    return tbase, ttrain, tvalidation, scaler

## FIASS Tuning

### IndexLSH

important_feautres = ['18', '12', '63', '62', '49', '2', '24', '19', '27', '64', '4',
       '47', '30', '9', '31', '23', '26', '50', '51', '60', '28', '67',
       '55', '69', '29', '53', '39', '11', '22', '38', '7', '15', '13',
       '14', '8', '32', '56', '58', '46', '41', '3', '52', '40', '20',
       '10', '66', '48', '36', '70', '0', '34', '37', '17', '71', '68',
       '43', '1', '25', '5', '16', '45', '42', '57', '54', '61', '35',
       '21', '33', '44', '65', '6', '59']

important_feautres_for_train = important_feautres.copy()
important_feautres_for_train.append("Target")

rscaler_base, rscaler_train, rscaler_validation, _ = checkRobustScaler(df_base[important_feautres], df_train[important_feautres_for_train], df_validation[important_feautres])

def train_index(df_base, df_train, short=True, k_value=50):
    start_train = time.time()
    dims = df_base.shape[1]

    idx_l2 = faiss.index_factory(dims, "Pad70,ITQ70,LSHrt", faiss.METRIC_L2)
    headon_solution_df_train = df_train.copy()

    df_base_sample = df_base.copy()

    if short:
        df_base_sample = df_base[:50000]
        headon_solution_df_train = headon_solution_df_train[headon_solution_df_train["Target"].isin(df_base_sample.index)]

        idx_l2.train(np.ascontiguousarray(df_base_sample.values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base_sample.values).astype('float32'))
    else:
        idx_l2.train(np.ascontiguousarray(df_base.values).astype('float32'))
        idx_l2.add(np.ascontiguousarray(df_base.values).astype('float32'))

    #base_index = {k: v for k, v in enumerate(df_base.index.to_list())}
    base_index = np.array(df_base.index.to_list())

    print("Learning index for:", time.time() - start_train)
    return df_base_sample, headon_solution_df_train, idx_l2, base_index
    

def searchByIndex(index, train, k_value):
    targets = train["Target"]
    train.drop("Target", axis=1, inplace=True)

    search_start = time.time()
    k=k_value
    vecs, idx = index.search(np.ascontiguousarray(train.values).astype('float32'), k)
    search_time = time.time() - search_start
    print("Search for:", search_time)
    return vecs, idx, search_time

def buildA5ScoreV2(df_base, df_train, short=True, k_value=50):
    base, train, index, base_index = train_index(df_base, df_train, short, k_value)
    targets = train["Target"]
    
    vecs, idx, search_time = searchByIndex(index, train, k_value)
    score = getAScore(targets.values.tolist(), idx.tolist(), base_index)

    searched_targets = []
    for i in idx:
        searched_targets.append([base_index[k] for k in i])
    
    return score, index, searched_targets, (k_value, search_time, score), base_index

%%time
results = []
#for kvalue in tqdm(list(range(5, 100, 5)) + list(range(100, 1000, 100)) + list(range(1000, 30000, 1000))):
for kvalue in tqdm(list(range(5, 100, 5)) + list(range(100, 1000, 100))):
    lsh_score, lsh_index, targets, stats, _ = buildA5ScoreV2(rscaler_base, rscaler_train, short=True, k_value=kvalue)
    results.append(stats)

[i for i in results if i[1] < 1.1 ]

if False:
    k_value = 4000
    lsh_score, lsh_index, searched_targets, full_stats, _= buildA5ScoreV2(rscaler_base, rscaler_train, short=False, k_value=k_value)
    print(f"Score on full dataset and k_value {k_value}", lsh_score)

#rscaler_train_top_100 = rscaler_train.sample(1000)
rscaler_train_top_100 = rscaler_train
lsh_score, lsh_index, target_index, stats, common_base_index = buildA5ScoreV2(rscaler_base, rscaler_train_top_100, short=False, k_value=200)
stats

if False:
    def defineMostImportantCandidatesForSingleRow(base, target, target_index):
        searched = base.loc[target_index]
        cosine_distances_v = cosine_similarity(searched.values, [target[searched.columns].values])
        cosine_distances_v = [item for sublist in cosine_distances_v for item in sublist]
    
        indexes = np.argpartition(cosine_distances_v, 100)[:-100]
        #TODO: Check it
        #indexes = np.argmin(cosine_distances_v)
        return searched.iloc[indexes].index.tolist()
    
    index_results = []
    for irow in trange(0, len(rscaler_train_top_100)):
        row = rscaler_train_top_100.iloc[irow]
        row_index = target_index[irow]
        index_results.append(defineMostImportantCandidatesForSingleRow(rscaler_base, row, row_index))

    acc = 0
    for irow in trange(0, len(rscaler_train_top_100)):
        if rscaler_train_top_100.iloc[irow].Target in index_results[irow]:
            acc+=1
    
    print(acc/len(rscaler_train_top_100) * 100)

#newTarget

### Validation predict

#validation_vecs, validation_idx, validation_search_time = searchByIndex(index, train, k_value)
validation_vect, validation_ids = lsh_index.search(np.ascontiguousarray(rscaler_validation.values).astype('float32'), 200)

searched_validation_targets = []
for i in validation_ids:
    searched_validation_targets.append([common_base_index[k] for k in i])

## Save searched indexes 

target_idx = rscaler_train_top_100.copy().reset_index()[["Id", "Target"]]
predictions = pd.DataFrame(pd.Series(target_index), columns=["predictions"])
train_results = pd.merge(target_idx, predictions, left_index=True, right_index=True)
train_results["is_match"] = train_results.apply(lambda x: x["Target"] in x["predictions"], axis=1)
train_results.set_index("Id", inplace = True)

#validation_predictions = pd.DataFrame(searched_validation_targets, columns=["predictions"])
validation_idx = df_validation.copy().reset_index()[["Id"]]
validation_predictions = pd.DataFrame(pd.Series(searched_validation_targets), columns=["predictions"])
validation_results = pd.merge(validation_idx, validation_predictions, left_index=True, right_index=True)
validation_results.head()
validation_results.set_index("Id", inplace = True)

df_validation

train_results.to_parquet("../tmp/predicted_targets.parquet", compression="gzip")
validation_results.to_parquet("../tmp/predicted_validations.parquet", compression="gzip")
faiss.write_index(lsh_index, "../tmp/lsh_search_index_2.faiss")
