==> notebook/8_Ranging.ipynb <==
print("Hello, world")

import numpy as np
import pandas as pd

from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import RobustScaler
from catboost import CatBoostClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier

from sklearn.metrics import accuracy_score, roc_curve, ConfusionMatrixDisplay
import matplotlib.pyplot as plt

## Предварительная настройка

df_base = pd.read_parquet("../tmp/cleared_df_base_new_features.parquet")
df_train = pd.read_parquet("../tmp/cleared_df_train_new_features.parquet")
df_validation = pd.read_parquet("../tmp/cleared_df_validation_new_features.parquet")
df_validation_target = pd.read_csv("../data/validation_answer.csv")
df_validation_target.set_index("Id", inplace=True)
df_ptargets = pd.read_parquet("../tmp/predicted_targets.parquet")
df_validation_predicts = pd.read_parquet("../tmp/predicted_validations.parquet")




def checkRobustScaler(base, train, validation):
    scaler = RobustScaler()
    scaler.fit(base)

    tbase = scaler.transform(base)
    tbase = pd.DataFrame(tbase, columns=base.columns)
    tbase["Id"] = base.index
    tbase = tbase.set_index(["Id"])
    
    ttrain = scaler.transform(train[base.columns])
    ttrain = pd.DataFrame(ttrain, columns=base.columns)
    ttrain["Id"] = train.index
    ttrain = ttrain.set_index(["Id"])
    ttrain["Target"] = train["Target"]

    tvalidation = scaler.transform(validation[base.columns])
    tvalidation = pd.DataFrame(tvalidation, columns=base.columns)
    tvalidation["Id"] = validation.index
    tvalidation = tvalidation.set_index(["Id"])

    return tbase, ttrain, tvalidation, scaler

df_base, df_train, df_validation, _ = checkRobustScaler(df_base, df_train, df_validation)

## Ранжирование
### Формирование тренировочной выборки

Первым делом необходимо составить новый датафрейм состоящий из двух половин:

1) Искомые векторы
2) Несколько кандидатов на сопоставление для каждого искомого вектора.

Последним столбцом для этого вектора станет matching содержащий правильный ответ: 1 - вектор подходит, 0 - вектор не подходим.

exploded_predictions = df_ptargets["predictions"].explode()

prepare_train = pd.merge(df_ptargets, exploded_predictions, left_index=True, right_index=True)[["Target", "predictions_y"]]
prepare_train = prepare_train.rename(columns={"predictions_y":"predictions"})
prepare_train["is_match"] = prepare_train.apply(lambda x: int(x["Target"] == x["predictions"]), axis=1)
prepare_train.head(10)

df_ranging = prepare_train.copy()
df_ranging.pop("Target")

df_ranging_with_train = pd.merge(df_ranging, df_train, right_index=True, left_index=True)
df_ranging_with_train.head()

df_ranging_with_train = pd.merge(df_ranging_with_train, df_base, left_on="predictions", right_index=True, suffixes=('',"_base"))

temp = df_ranging_with_train.copy()
df_ranging_target = temp.pop("is_match")
df_ranging_features = temp.drop(columns=["predictions", "Target"])

df_ranging_target = df_ranging_target.astype('int8')

df_ranging_target

del df_ranging_with_train

del df_ranging

del df_train

del prepare_train

### Validation dataset

exploded_predictions_validation = df_validation_predicts.copy()["predictions"].explode()
exploded_predictions_validation

df_validation

validation_with_predict = pd.merge(df_validation, exploded_predictions_validation, left_index=True, right_index=True)
validation_with_predict = pd.merge(validation_with_predict, df_base, left_on="predictions", right_index=True, suffixes=('',"_base"))
validation_with_predict = pd.merge(validation_with_predict, df_validation_target, left_index=True, right_index=True)
validation_with_predict[["predictions", "Expected"]]

validation_with_predict["is_matching"] = validation_with_predict.apply(lambda x: int(x["Expected"] == x["predictions"]), axis=1)
validation_with_predict["is_matching"] = validation_with_predict["is_matching"].astype('int8')
validation_with_predict.info()

#temp = validation_with_predict.copy()
validation_with_predict_target = validation_with_predict.pop("is_matching")
validation_with_predict_features = validation_with_predict.drop(columns=["predictions", "Expected"])

validation_with_predict

## Логистическая регрессия

def checkModel(model, train_features, train_target, valid_feauters, valid_target):
    model.fit(train_features, train_target)

    valid_predictions = model.predict(valid_feauters)
    valid_probabilities = model.predict_proba(valid_feauters)[:, 1]
    valid_accuracy = accuracy_score(valid_target, valid_predictions)
    
    print('valid_accuracy: {:.2%}'.format(valid_accuracy))
    ConfusionMatrixDisplay.from_estimator(model, valid_feauters, valid_target, cmap=plt.cm.Blues, normalize='true'); # матрица ошибок

    # ROC-кривая
    fpr, tpr, thresholds = roc_curve(valid_target, valid_probabilities)
    
    plt.figure(figsize = (6, 6))
    plt.step(fpr, tpr, where='post')
    plt.xlim(0, 1)
    plt.ylim(0, 1)
    plt.xlabel('FPR')
    plt.ylabel('TPR')
    plt.title('ROC-кривая')
    plt.plot([0, 1], [0, 1], linestyle='--')
    plt.grid(True)
    plt.show()

    return valid_probabilities


df_ranging_features.info()

df_ranging_target.info()

logreg_valid_probabilities = checkModel(LogisticRegression(solver="liblinear", class_weight='balanced', random_state=42, C=1.5), df_ranging_features, df_ranging_target, validation_with_predict_features, validation_with_predict_target)

print(logreg_valid_probabilities)

logreg_valid_proba_series = pd.Series(logreg_valid_probabilities)

logres_valid_candidates = []
for i in range(0, len(logreg_valid_proba_series), 100):
    query = logreg_valid_proba_series[i : i+100]
    index = query.sort_values(ascending=False)[0:5].index
    logres_valid_candidates.append(index)

# контроль
print('Кол-во предсказаний:', len(logres_valid_candidates))
print('Кол-во кандидатов:', len(logres_valid_candidates[0]))


### ---------------------------------------------------------------------------------------------------------

acc = 0
for target, candidates in zip(df_validation_target.values.tolist(), logres_valid_candidates):
    acc += int(target in validation_with_predict.iloc[candidates.tolist()]["predictions"].values)

print('valid_accuracy@{}: {:.2f}%'.format(5, acc / len(logres_valid_candidates) * 100))


acc

### Catboost

cbc = CatBoostClassifier(auto_class_weights = 'Balanced', verbose=100, random_state=42, n_estimators=1000, bootstrap_type="Bernoulli")

cbc.fit(df_ranging_features.values, df_ranging_target.values)

predictions = cbc.predict_proba(validation_with_predict_features.values)

df_predictions = pd.DataFrame(predictions)
df_predictions.drop([0], axis=1, inplace=True)
df_predictions.columns=['predictions']

cbc_valid_proba_series = pd.Series(df_predictions['predictions'])

cbc_valid_candidates = []
for i in range(0, len(cbc_valid_proba_series), 200):
    query = cbc_valid_proba_series[i : i+200]
    index = query.sort_values(ascending=False)[0:5].index
    cbc_valid_candidates.append(index)

# контроль
print('Кол-во предсказаний:', len(cbc_valid_candidates))
print('Кол-во кандидатов:', len(cbc_valid_candidates[0]))


acc = 0
for target, candidates in zip(df_validation_target.values.tolist(), cbc_valid_candidates):
    acc += int(target in validation_with_predict.iloc[candidates.tolist()]["predictions"].values)

print('valid_accuracy@{}: {:.2f}%'.format(5, acc / len(cbc_valid_candidates) * 100))
